refactor(packages): apply ts naming guideline to utility files (#728)
