+++ packages/ui/treeconsole/base/src/adapters/commands/DraftCommands.ts
@@ -15,7 +15,7 @@ import type {
   TreeNode,
 } from '@hierarchidb/tree-api';
 import type { WorkerAPI } from '@hierarchidb/worker-api';
-import { createCommand } from '~/adapters/utils';
+import { createCommand } from '~/adapters/commandEnvelopeFactories';
 import type { CommandAdapterOptions } from '~/types/index';
 import { TreeConsoleAdapterError } from '~/types/index';
