+++ packages/ui/treeconsole/base/src/adapters/commands/TreeMutationCommands.ts
@@ -12,7 +12,7 @@ import type {
   RestoreFromArchivePayload,
 } from '@hierarchidb/tree-api';
 import type { WorkerAPI } from '@hierarchidb/worker-api';
-import { createCommand } from '~/adapters/utils';
+import { createCommand } from '~/adapters/commandEnvelopeFactories';
 import type { CommandAdapterOptions } from '~/types/index';
 import { TreeConsoleAdapterError } from '~/types/index';
