+++ packages/ui/treeconsole/base/src/adapters/subscriptions/TreeObservableAdapter.ts
@@ -11,7 +11,7 @@ import type { NodeId } from '@hierarchidb/core-types';
 import type { TreeNodeEvent } from '@hierarchidb/tree-api';
 import type { AdapterContext, UnsubscribeFunction } from '~/types/index';
 import { TreeConsoleAdapterError } from '~/types/index';
-import { createCommand } from '~/adapters/utils';
+import { createCommand } from '~/adapters/commandEnvelopeFactories';
 import { sanitizeForComlink } from './comlinkSanitizer';
 
 type TreeNodeEventCallback = (event: TreeNodeEvent) => void;
